Reorganização -"gerar_rodadas" foi movido para o arquivo "functions.py". -Adicionado a função "tempo_execução", responsável por calcular o tempo  de execução do código.
